Refactor test file locations and package names
Renamed goja_nbt_test.go to jsbindings/goja_nbt_test.go and nbt_test.go to nbt/nbt_test.go. Updated package declarations to match new directories for better organization and clarity.
