+++ cmd/tests/nbt/nbt_test.go
@@ -1,4 +1,4 @@
-package tests
+package nbt
 
 import (
 	"bytes"
